Show file availability in percents

+++ client/src/logic/model.scala
@@ -23,7 +23,7 @@ object RootModel {
 case class TorrentModel(
   infoHash: InfoHash,
   connected: Int,
-  availability: List[Int],
+  availability: List[Double],
   metadata: Option[Either[String, Metadata]]
 ) {
   def withMetadata(metadata: Metadata): TorrentModel = copy(metadata = Some(Right(metadata)))
